# apps/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  apikey.APIKey:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      modelId:
        type: string
      providerId:
        type: string
      service:
        type: string
      systemPrompt:
        type: string
      teamId:
        type: string
      updatedAt:
        type: string
    type: object
  apikey.CreateAPIKeyRequest:
    properties:
      key:
        type: string
      modelId:
        type: string
      providerId:
        type: string
      service:
        type: string
      systemPrompt:
        type: string
    type: object
  apikey.UpdateAPIKeyRequest:
    properties:
      isActive:
        type: boolean
      key:
        type: string
      modelId:
        type: string
      providerId:
        type: string
      service:
        type: string
      systemPrompt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@seo-backend.com
    name: API Support
  description: API Documentation for SEO Backend Service
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: SEO Backend API
  version: "1.0"
paths:
...
  /api/api-keys:
    get:
      consumes:
      - application/json
      description: Get all API keys for the current user/team
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            items:
              $ref: '#/definitions/apikey.APIKey'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all API keys
      tags:
      - API Keys
    post:
      consumes:
      - application/json
      description: Create a new API key for accessing AI models
      parameters:
      - description: API key creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apikey.CreateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 'id: key_id, message: created successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request - missing required fields (service, provider_id,
            model_id, api_key)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new API key
      tags:
      - API Keys
  /api/api-keys/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an API key by ID
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'id: key_id, message: deleted successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: API key not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete an API key
      tags:
      - API Keys
    get:
      consumes:
      - application/json
      description: Get a specific API key by its ID
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/apikey.APIKey'
        "404":
          description: API key not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get API key by ID
      tags:
      - API Keys
    put:
      consumes:
      - application/json
      description: Update an existing API key by ID
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      - description: API key update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apikey.UpdateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'id: key_id, message: updated successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: API key not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an API key
      tags:
      - API Keys
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"